Refactor files a little
