--- dask_databricks/databricks/__init__.py
@@ -1,1 +1,0 @@
-from .databrickscluster import DatabricksCluster
