--- dask_databricks/tests/__init__.py
@@ -1,4 +1,0 @@
-# SPDX-FileCopyrightText: 2023-present Dask Developers
-#
-# SPDX-License-Identifier: BSD-3
-
