+++ dask_databricks/tests/test_databricks.py
@@ -5,7 +5,7 @@ from dask.distributed import Client
 from distributed.deploy import Cluster, LocalCluster
 
 
-from dask_databricks import databricks
+from dask_databricks import DatabricksCluster
 
 @pytest.fixture(scope="session")
 def dask_cluster():
@@ -33,15 +33,15 @@ def set_spark_local_ip():
 
 def test_databricks_cluster_raises_key_error_when_initialised_outside_of_databricks(remove_spark_local_ip):
     with pytest.raises(KeyError):
-        databricks.DatabricksCluster()
+        DatabricksCluster()
 
 def test_databricks_cluster_create(set_spark_local_ip, dask_cluster):
-    cluster = databricks.DatabricksCluster()
+    cluster = DatabricksCluster()
     assert isinstance(cluster, Cluster)
 
 
 def test_databricks_cluster_create_client(set_spark_local_ip, dask_cluster):
-    cluster = databricks.DatabricksCluster()
+    cluster = DatabricksCluster()
     client = Client(cluster)
     assert isinstance(client, Client)
     assert client.submit(sum, (10, 1)).result() == 11
